Add files via upload
This program takes number of persons and amounts taken by each other as input, then minimizes the total cashflow.

Using the Greedy technique, all amounts are settled for one individual and repeated for the remaining n-1 persons.
To select the first person, compute the net amount for each person, which is derived by subtracting all amounts to pay from all amounts to be paid. The majority of creditors and debtors are determined after calculating the net amount for each individual. The person with a minimum of two is the first to be resolved and removed from the list. Let x be the least of the two amounts. We pay 'x' amount from maximum debtor to maximum creditor and settle with one individual. If x equals the maximum debit, the maximum debtor is settled; otherwise, the maximum creditor is settled.